+++ server/database/migrations/2021_06_22_145642_create_purchase_payments_table.php
@@ -18,7 +18,7 @@ class CreatePurchasePaymentsTable extends Migration
             $table->id();
             $table->string('reference')->default("INV/PR_1110");
 
-            $table->bigInteger('amount');
+            $table->float('amount', 10, 0);
             $table->bigInteger('payment_method')->default(Constants::PAYMENT_CASH);
             $table->text('note')->nullable();
