+++ server/database/migrations/2021_07_13_154848_create_purchase_return_payments_table.php
@@ -18,8 +18,8 @@ class CreatePurchaseReturnPaymentsTable extends Migration
             $table->id();
             $table->string('reference')->default("INV/RT_1110");
 
-            $table->bigInteger('amount');
-            $table->bigInteger('payment_method')->default(Constants::PAYMENT_CASH);
+            $table->float('amount', 10, 0);
+            $table->integer('payment_method')->default(Constants::PAYMENT_CASH);
             $table->text('note')->nullable();
 
             $table->unsignedBigInteger('purchase_return_id');
