+++ server/database/migrations/2021_07_17_196930_create_sale_returns_table.php
@@ -18,14 +18,14 @@ class CreateSaleReturnsTable extends Migration
             $table->id();
             $table->string('reference')->default("RT_1110");
 
-            $table->integer('tax')->default(0);
-            $table->integer('discount')->default(0);
+            $table->float('tax', 10, 0)->default(0);
+            $table->float('discount', 10, 0)->default(0);
             $table->tinyInteger("discount_method")->default(Constants::DISCOUNT_FIXED); // 0 Fixed, 1 Percent
             $table->tinyInteger('status')->default(Constants::SALE_RETURN_RECEIVED);
             $table->tinyInteger('payment_status')->default(Constants::PAYMENT_STATUS_UNPAID);
-            $table->integer('shipping')->default(0);
-            $table->float('total_price')->default(0);
-            $table->float('paid')->default(0);
+            $table->float('shipping', 10, 0)->default(0);
+            $table->float('total_price', 10, 0)->default(0);
+            $table->float('paid', 10, 0)->default(0);
 
             $table->text('note')->nullable();
             $table->date('date')->default(date('Y-m-d'));
