+++ server/database/migrations/2021_07_17_197544_create_sale_return_details_table.php
@@ -17,11 +17,11 @@ class CreateSaleReturnDetailsTable extends Migration
         Schema::create('sale_return_details', function (Blueprint $table) {
             $table->id();
 
-            $table->integer("price");
-            $table->float("quantity");
-            $table->integer('tax')->default(0);
+            $table->float("price", 10, 0);
+            $table->float("quantity", 10, 0);
+            $table->float('tax', 10, 0)->default(0);
             $table->tinyInteger("tax_method")->default(Constants::TAX_EXCLUSIVE); // 0 Exclusive, 1 Inclusive
-            $table->integer('discount')->default(0);
+            $table->float('discount', 10, 0)->default(0);
             $table->tinyInteger("discount_method")->default(Constants::DISCOUNT_FIXED); // 0 Fixed, 1 Percent
 
             $table->bigInteger('variant_id')->nullable();
