+++ server/database/migrations/2021_07_25_080249_create_transfers_table.php
@@ -18,13 +18,13 @@ class CreateTransfersTable extends Migration
             $table->id();
             $table->string('reference')->default("TR_1110");
 
-            $table->integer('tax')->default(0);
-            $table->integer('discount')->default(0);
-            $table->integer('shipping')->default(0);
+            $table->float('tax', 10, 0)->default(0);
+            $table->float('discount', 10, 0)->default(0);
+            $table->float('shipping', 10, 0)->default(0);
             $table->tinyInteger("discount_method")->default(Constants::DISCOUNT_FIXED); // 0 Fixed, 1 Percent
-            $table->float('total_price')->default(0);
+            $table->float('total_price', 10, 0)->default(0);
 
-            $table->float('items_count');
+            $table->float('items_count', 10, 0);
 
             $table->text('note')->nullable();
             $table->date('date')->default(date('Y-m-d'));
